requirements_dev.txt
-r requirements.txt
pip
bumpversion
wheel
watchdog
flake8
tox
coverage
cryptography
PyYAML
pytest
pytest-runner
pytest-cov
Sphinx
sphinx_rtd_theme
nbsphinx
numpydoc
flask
